It's July 1st, Haejin's birthday
She's never gonna see this stuff but I hope she had a great day today hehe

We rowed a boat for the first time together at Central Park, walked around for a bit before deciding to head home and chill for a bit, because the weather was disgusting outside and we had both had some terrible sleep. Ordered some food and ate lunch with my sister, who came over with balloons and some flowers for Haejin. She left, we got ready for dinner reservations at Barano in Williamsburg. Before taking the bus, made a new friend with a guy named Charles, who was helping us look out for the bus. We got him a drink at a bubble tea shop and talked to him for a bit, was really sweet. Got off, ordered Barano's meatballs, Nonna pizza, and linguine pasta. Finished eating, took home the leftover pizza, and walked down to the waterside to enjoy the sunset. Took some time to relax and talk and took the ferry together back to LIC. Stayed at Gantry for a little bit more to enjoy the breeze before calling an Uber to head home.

Wasn't the craziest day in terms of activities, but knocked off a lot of water-related things that she has always wanted to do since she got to the city. So that's a win in itself.

Ooh, forgot to mention. We ordered and picked up cake from Billy's Bakery in Chelsea yesterday and when we got home, we had some cake using this big ass pasta fork that we bought for 30 dollars a year ago. Now chilling, writing a commit on GitHub about the day I had with my girlfriend.

If you ever do see this, for whatever reason, know that I love you so much, and I hope we get to enjoy the rest of your birthday weekend together, doing whatever it is that makes you happy.

BIG 26git add client/tsconfig.json !